fix: ignore team requested reviews
The GraphQL request didn't return all of the informations required to distinguish users & teams because of OAuth App access restrictions on the Angular repository.
I switched back to rest requests which now contain all of the information necessary (that was not the case a few months ago).
We now only count the user request reviews instead of users + teams.