perf: cache config file in firebase
Cache the config files for each repo in Firebase instead of requesting them from Github for each PR.
Whenever the config is needed, we get it from Firebase, or from Github if we don't have the config in cache (in which case we cache it at the same time).
Whenever there is a push commit to master, we check if the config file was updated, and if that's the case we update the cached config.
There is also a new get function that you can call to delete all of the cache. This can be used to force a refresh of the cache if the cached config is wrong (if we missed the push event for example).