fix: add g3status if enabled & missing
When the synchronize event is received after the PR created event, we are unable to update the PR with the g3 status because we don't know yet to which PR this commit belongs.
Now we will add the g3 status if it's missing (and g3 status is enabled for this repository) whenever a status update is requested.
This mitigates the Github events timing issues.