fix: don't update g3status on merge label
We were updating the g3status when the merge label was added. This meant that if a g3sync had updated the status, it was overwritten.
With this PR we now only update the g3status on sync (new commit or force push) or when the g3status is missing.